Separate the base, handler and message
The shared methods in SimpleFileTransferServer and
SimpleFileTransferClient are moved to SimpleFileTransferBase.
The server handler and client handler are gathered into
handler.py while the protocol message definitions are moved
to message.py

The executable are extracted to client.py and server.py in
the outer folder to avoid executing from withing the package

Currently, only the directory listing function is implemented.