Actual separation of Client, Server, and Base
In the previous commit (d627aec3), I've done the separation of
Client, Server and Base. But didn't actually remove the duplicated
code from the client and server, which result in the methods in
Base not getting called when it should. Also, the direct inheritance
to Protocol is removed in Client and Server. Now only the Base inherits
the Protocol.